+++ includes/providers/gam/api/class-api.php
@@ -129,26 +129,22 @@ class Api {
 	/**
 	 * Get a WP_Error object from an optional ApiException or message.
 	 *
-	 * @param ApiException $exception       Optional Google Ads API exception.
-	 * @param string       $default_message Optional default message to use.
+	 * @param \Exception|ApiException $exception       Optional PHP exception or Google API exception.
+	 * @param string                  $default_message Optional default message to use.
 	 *
 	 * @return WP_Error
 	 */
-	public function get_error( ApiException $exception = null, $default_message = null ) {
+	public function get_error( $exception = null, $default_message = null ) {
 		$error_message = $default_message;
 		$errors        = [];
-		if ( ! is_null( $exception ) ) {
+		if ( ! is_null( $exception ) && $exception instanceof ApiException ) {
 			$error_message = $error_message ?? $exception->getMessage();
 			foreach ( $exception->getErrors() as $error ) {
 				$errors[] = $error->getErrorString();
 			}
 		}
-		try {
-			$network_code = $this->get_network_code();
-		} catch ( \Exception $e ) {
-			$network_code = null;
-		}
-		$message_map = [
+		$network_code = $this->network_code;
+		$message_map  = [
 			'UniqueError.NOT_UNIQUE'                    => __( 'Name must be unique.', 'newspack-ads' ),
 			'CommonError.CONCURRENT_MODIFICATION'       => __( 'Unexpected API error, please try again in 30 seconds.', 'newspack-ads' ),
 			'PermissionError.PERMISSION_DENIED'         => __( 'You do not have permission to perform this action. Make sure to connect an account with administrative access.', 'newspack-ads' ),
@@ -207,16 +203,16 @@ class Api {
 				'applicationName' => self::APP,
 			],
 		];
-		/** If a network code is not yet available, use first from list. */
-		if ( ! $this->network_code ) {
-			$session  = ( new AdManagerSessionBuilder() )->from( new Configuration( $config ) )->withOAuth2Credential( $this->credentials )->build();
-			$networks = $this->get_networks( $session );
-			if ( ! empty( $networks ) ) {
-				$this->network_code = $networks[0]->getNetworkCode();
-			}
+
+		// Get network code and add it to the session config.
+		try {
+			// We're silencing errors here because the SDK throws a warning when the request config doesn't include a network code.
+			$config['AD_MANAGER']['networkCode'] = @$this->get_network_code( ( new AdManagerSessionBuilder() )->from( new Configuration( $config ) )->withOAuth2Credential( $this->credentials )->build() ); // phpcs:ignore WordPress.PHP.NoSilencedErrors.Discouraged
+		} catch ( \Exception | \ApiException $e ) {
+			return $this->get_error( $e, __( 'Unable to fetch network code from GAM.', 'newspack-ads' ) );
 		}
-		$config['AD_MANAGER']['networkCode'] = $this->network_code;
-		$this->session                       = ( new AdManagerSessionBuilder() )->from( new Configuration( $config ) )->withOAuth2Credential( $this->credentials )->build();
+		// Generate and return session.
+		$this->session = ( new AdManagerSessionBuilder() )->from( new Configuration( $config ) )->withOAuth2Credential( $this->credentials )->build();
 		return $this->session;
 	}
 
@@ -269,12 +265,14 @@ class Api {
 	/**
 	 * Get user's GAM network. Defaults to the first found network if not found or empty.
 	 *
+	 * @param AdManagerSession $session Optional session to use.
+	 *
 	 * @return Network GAM network.
 	 *
 	 * @throws \Exception If there is no GAM network to use.
 	 */
-	public function get_network() {
-		$networks     = $this->get_networks();
+	public function get_network( $session = null ) {
+		$networks     = $this->get_networks( $session );
 		$network_code = $this->network_code;
 		if ( empty( $networks ) ) {
 			throw new \Exception( __( 'Missing GAM Ad network.', 'newspack-ads' ) );
@@ -292,10 +290,12 @@ class Api {
 	/**
 	 * Get user's GAM network code.
 	 *
+	 * @param AdManagerSession $session Optional session to use.
+	 *
 	 * @return int GAM network code.
 	 */
-	public function get_network_code() {
-		return $this->get_network()->getNetworkCode();
+	public function get_network_code( $session = null ) {
+		return $this->get_network( $session )->getNetworkCode();
 	}
 
 	/**