+++ newspack-ads.php
@@ -5,7 +5,7 @@
  * Description:     Ad services integration.
  * Author:          Automattic
  * License:         GPL2
- * Version:         1.44.0
+ * Version:         1.44.5
  *
  * @package         Newspack
  */